# docs/static/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: eInvoice Storage - OpenAPI 3.0
  description: |-
    A description of the Invoicer REST API
  contact:
    email: patricklachlanbell@gmail.com
  version: 1.0.1
...
servers:
  - url: http://node-express-env.eba-9atmj7qr.ap-southeast-2.elasticbeanstalk.com
    description: Production server (uses live data and can be viewed on frontend)
paths:
...
  /user/signup:
    post:
      tags:
        - user
      summary: Sign up for a new account.
      description: Creates a new account with the given details.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - firstname
                - lastname
                - email
              properties:
                username:
                  type: string
                  example: "johndoe"
                firstname:
                  type: string
                  example: "John"
                lastname:
                  type: string
                  example: "Doe"
                password:
                  type: string
                  description: If you are not using HTTPS ensure this password can not be intercepted!
                  example: "XM92eJEFHFTt214v2eetG4"
                email:
                  type: string
                  description: Must be valid format for an email.
                  example: johndoe@example.com
        required: true
      responses:
        '200':
          description: Details of account which was created.
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                  - user
                properties:
                  message:
                    type: string
                    example: "Signup successful"
                  user:
                    type: object
                    required:
                      - username
                      - _id
                    properties:
                      username:
                        type: string
                        example: "johndoe"
                      _id:
                        type: string
                        description: objectId of new user
                        example: "64226af987d1e2174b494491"
        '400':
          $ref: '#/components/responses/InvalidError'
        '500':
          description: Internal server error
  /user/login:
    post:
      tags:
        - user
      summary: Login to an existing user's account.
      description: Checks that the user with the given username has a matching password then returns a JWT authorization token for a new session.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                api:
                  type: boolean
                  example: true
                  description: Token will never expire if api is true.
                username:
                  type: string
                  example: "johndoe"
                password:
                  description: If you are not using HTTPS ensure this password can not be intercepted!
                  type: string
                  example: "XM92eJEFHFTt214v2eetG4"
        required: true
      responses:
        '200':
          description: Details of newly authenticated session.
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                  - token
                properties:
                  message:
                    type: string
                    example: "Login successful"
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjp7ImlkIjoiNjQyMTQwMzU4N2QxZTIxNzRiNDk0NDhkIiwidXNlcm5hbWUiOiJwYXRyaWNrIn0sImlhdCI6MTY4MDI2NTE2NiwiZXhwIjoxNjgwMjY4NzY2fQ.ecnLCgXQ-JufzFXK7RYl23bQehT3cakZj7Tmd7pwd8A"
                    description: JSON Web Token, add this to headers as a bearer token to authorize future requests. This will expire after a certain time.
        '400':
          $ref: '#/components/responses/InvalidError'
        '401':
          description: Message describing why the user was not authenticated.
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                properties:
                  message:
                    type: string
                    example: "Incorrect username or password"
        '500':
          description: Internal server error
  /user/{id}:
    get:
      tags:
        - user
      summary: Retrieve a user with the specified id or the authenticating user.
      description: The specified user id or authenticating user's id is searched and if it is found a description of the user is returned. Does not require authorization, but if authorized the user information returned could be more detailed.
      parameters:
        - in: path
          name: id
          description: objectId of the user to be retrieved. If not provided then request must be authenticated and the authenticating user's id is used.
          required: false
          schema:
            type: string
            minLength: 24
            maxLength: 24
            example: "6421d1e887d1e2174b494492"
      responses:
        '200':
          description: User details, depending on the authorization a different amount of information is exposed.
          content:
            application/json:
              schema:
                type: object
                description: If the authorizing user's id does not have permission then only 'firstname', 'lastname', 'username' and 'dateCreated' is sent in 'user'.
                required:
                  - found
                  - user
                properties:
                  found:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/InvalidError'
        '404':
          description: Acknowledgement that no user with the specified id exists.
          content:
            application/json:
              schema:
                type: object
                required:
                  - found
                properties:
                  found:
                    type: boolean
                    example: false
        '500':
          description: Internal server error
      security:
        - bearerAuth:
            - read:user
    patch:
      tags:
          - user
      summary: Change the user with the specified id or the authenticating user's details.
      description: The user with specified id or authenticating user is modified according to body of request.
      parameters:
        - name: id
          in: path
          description: objectId of the user to be retrieved. If not provided then request must be authenticated and the authenticating user's id is used.
          required: false
          schema:
            type: string
            minLength: 24
            maxLength: 24
            example: "6421d1e887d1e2174b494492"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: "janedoe"
                firstname:
                  type: string
                  example: "Jane"
                lastname:
                  type: string
                  example: "Doe"
                email:
                  type: string
                  description: Must be valid format for an email.
                  example: janedoe@example.com
        required: false
      responses:
        '200':
          description: The user with the given id was successfully modified.
          content:
            application/json:
              schema:
                type: object
                required:
                  - acknowledged
                properties:
                  aknowledged:
                    type: boolean
                    example: true
                  matchedCount:
                    type: integer
                    example: 1
        '400':
          $ref: '#/components/responses/InvalidError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: A user with the given id doesn't exist.
          content:
            application/json:
              schema:
                type: object
                required:
                  - acknowledged
                properties:
                  aknowledged:
                    type: boolean
                    example: true
                  matchedCount:
                    type: integer
                    example: 0
        '500':
          description: Internal server error
      security:
        - bearerAuth:
            - edit:user
    delete:
      tags:
        - user
      summary: Delete the user with the specified id or the authenticating user.
      description: Delete the user with the specified id or the authenticating user if the authorizing user has permission.
      parameters:
        - name: id
          in: path
          description: objectId of the user to be deleted. If not provided then request must be authenticated and the authenticating user's id is used.
          required: false
          schema:
            type: string
            minLength: 24
            maxLength: 24
            example: "6421d1e887d1e2174b494492"
      responses:
        '200':
          description: User was found and deleted.
          content:
            application/json:
              schema:
                type: object
                required:
                  - deletedCount
                  - acknowledged
                properties:
                  acknowledged:
                    type: boolean
                    example: true
                  deletedCount:
                    type: integer
                    example: 1
        '400':
          $ref: '#/components/responses/InvalidError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: A user with id doesn't exist or user is not authorized to delete it.
          content:
            application/json:
              schema:
                type: object
                required:
                  - deletedCount
                  - acknowledged
                properties:
                  acknowledged:
                    type: boolean
                    example: true
                  deletedCount:
                    type: integer
                    example: 0
        '500':
          description: Internal server error
      security:
        - bearerAuth:
            - write:user
...
components:
  schemas:
    User:
      type: object
      required:
        - username
        - firstname
        - lastname
        - email
        - hash
        - salt
        - creationDate
      properties:
        username:
          type: string
          example: "johndoe"
        firstname:
          type: string
          example: "John"
        lastname:
          type: string
          example: "Doe"
        email:
          type: string
          description: Must be valid format for an email.
          example: johndoe@example.com
        hash:
          type: string
          description: Salted hash based on the user's password
          example: "p7D6jYqPs6x3mL2CBH11Jt4sHOg72MjCNc0QtAwrujB="
        salt:
          type: string
          description: The salt used to hash the password
          example: "u14rCPA1gV3K6ghnr6HoOQ=="
        creationDate:
          example: "1975-08-19T23:15:30.000Z"
          description: Serialized JS Date object, see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/toJSON
          type: string
...
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid, possibly because it has expired.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
    InvalidError:
      description: Parameters don't match the schema (validated with [YUP](https://www.npmjs.com/package/yup)) or an internal server error was encountered.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: An optional message describing the error
                example: "field body.recipientName missing"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT